[libgum] Organise libgum files based on category.
Currently we have all the libgum files in lib/gum. This patch organises
the files based on the what they are catering to. We now have the
following libgum structure:
/lib/                 <-top-level directory
/lib/c                <- c-headers
/lib/gum              <- gum includes
/lib/gum/kernel/linux <- cogent implementation for linux kernel api
/lib/gum/fs/linux     <- cogent implementation for linux vfs api
/lib/gum/common       <- common cogent implemtations that are used across
/lib/gum/anti         <- antiquoted C implementations.
/lib/gum/test
