--- cogent/lib/gum/abstract/wordarray.ah
@@ -1,16 +1,0 @@
-/*
- * Copyright 2016, NICTA
- *
- * This software may be distributed and modified according to the terms of
- * the GNU General Public License version 2. Note that NO WARRANTY is provided.
- * See "LICENSE_GPLv2.txt" for details.
- *
- * @TAG(NICTA_GPL)
- */
-
-struct $id:(WordArray a) {
-	u32 len;
-	$ty:a* values;
-};
-
-typedef struct $id:(WordArray a) $id:(WordArray a);
