+++ cogent/lib/gum/anti/abstract/AllocPool.ah
