+++ cogent/lib/gum/anti/abstract/UArray.ah
