+++ cogent/lib/gum/common/array.cogent
