+++ cogent/lib/gum/common/buffer.cogent
