+++ cogent/lib/gum/common/common.cogent
