+++ cogent/lib/gum/common/crc32.cogent
