+++ cogent/lib/gum/common/iterator.cogent
