+++ cogent/lib/gum/common/log.cogent
