+++ cogent/lib/gum/common/wordarray.cogent
