+++ cogent/lib/gum/fs/linux/stat.cogent
