+++ cogent/lib/gum/kernel/linux/osbuffer.cogent
@@ -8,7 +8,7 @@
 -- @TAG(NICTA_GPL)
 --
 
-include "common.cogent"
+include "../../common/common.cogent"
 
 type OSBuffer
 type OSBufferOffset = U32
