+++ cogent/lib/gum/libgum_tc_test.sh
@@ -21,8 +21,9 @@ shopt -s globstar
 # since some libgum stuff depends on this type
 echo $'type FsInode\ntype FsState\n_COGENT_LOG_LEVEL: U32\n_COGENT_LOG_LEVEL = 0' > "$TEST_FILE"
 
+flist=`find . -name *.cogent`
 # include everything
-for src in *.cogent
+for src in $flist
 do
   echo "include \"$src\"" >> "$TEST_FILE"
 done
