+++ cogent/tests/antiquote-tests/test-array/array_simple.cogent
@@ -8,8 +8,8 @@
 -- @TAG(NICTA_GPL)
 --
 
-include <gum/array.cogent>
-include <gum/test.cogent>
+include <gum/common/array.cogent>
+include <gum/test/test.cogent>
 
 const_arraySize: U32
 const_arraySize = 4
