+++ cogent/tests/antiquote-tests/test-iterator/iterator_sum.cogent
@@ -8,7 +8,7 @@
 -- @TAG(NICTA_GPL)
 --
 
-include <gum/iterator.cogent>
+include <gum/common/iterator.cogent>
 
 seq32_generator: (#{acc: (U32, U32), obsv: U32!}) -> GeneratorResult U32 () () (U32, U32)
 seq32_generator (#{acc = (cur_iter, acc), obsv = max_iter}) =
