+++ cogent/tests/antiquote-tests/test-wordarray/wordarray_simple.cogent
@@ -8,8 +8,8 @@
 -- @TAG(NICTA_GPL)
 --
 
-include <gum/wordarray.cogent>
-include <gum/common.cogent>
+include <gum/common/wordarray.cogent>
+include <gum/common/common.cogent>
 
 const_arraySize: U32
 const_arraySize = 4
