+++ impl/bilby/cogent/plat/linux/wrapper.h
@@ -12,7 +12,7 @@
 # define _WRAP_LINUX_H_
 
 #include <adt.h>
-#include <lib/rbt.h>
+#include <lib/c/rbt.h>
 
 #define bilbyfs_assert(v) BUG_ON(!(v))
